+++ pkg/types/boxes/boxestype_impl.go
@@ -592,12 +592,11 @@ func (doc *BoxesDocument) drawCommentTextsCustomLabels(drawing types.Drawing, c 
 		if c.Text == "" {
 			continue
 		}
+		className := "comment"
 		if c.ConnectionIndex != nil {
-			className := fmt.Sprintf("connection conLine_%d", *c.ConnectionIndex)
-			drawing.DrawCircleWithBorderTextAndClass(c.Label, markerX, currentY, doc.CommentMarkerRadius, &c.Format.Fill, &c.Format.Line, &c.Format.FontMarker, className)
-		} else {
-			drawing.DrawCircleWithBorderAndText(c.Label, markerX, currentY, doc.CommentMarkerRadius, &c.Format.Fill, &c.Format.Line, &c.Format.FontMarker)
+			className = fmt.Sprintf("%s connection conLine_%d", className, *c.ConnectionIndex)
 		}
+		drawing.DrawCircleWithBorderTextAndClass(c.Label, markerX, currentY, doc.CommentMarkerRadius, &c.Format.Fill, &c.Format.Line, &c.Format.FontMarker, className)
 		c.Format.FontText.Anchor = types.FontDefAnchorEnum_left
 		c.Format.FontText.MaxLenBeforeBreak = doc.Boxes.Width
 		drawing.DrawText(c.Text, textX, currentY-(2*doc.CommentMarkerRadius), 0, &c.Format.FontText)
@@ -635,12 +634,11 @@ func (doc *BoxesDocument) drawCommentTextsStdLabels(currentY int, drawing types.
 		}
 		c.Format.FontText.Anchor = types.FontDefAnchorEnum_left
 		c.Format.FontText.MaxLenBeforeBreak = doc.Boxes.Width
+		className := "comment"
 		if c.ConnectionIndex != nil {
-			className := fmt.Sprintf("connection conLine_%d", *c.ConnectionIndex)
-			drawing.DrawCircleWithBorderTextAndClass(c.Label, markerX, currentY, doc.CommentMarkerRadius, &c.Format.Fill, &c.Format.Line, &c.Format.FontMarker, className)
-		} else {
-			drawing.DrawCircleWithBorderAndText(c.Label, markerX, currentY, doc.CommentMarkerRadius, &c.Format.Fill, &c.Format.Line, &c.Format.FontMarker)
+			className = fmt.Sprintf("%s connection conLine_%d", className, *c.ConnectionIndex)
 		}
+		drawing.DrawCircleWithBorderTextAndClass(c.Label, markerX, currentY, doc.CommentMarkerRadius, &c.Format.Fill, &c.Format.Line, &c.Format.FontMarker, className)
 		drawing.DrawText(c.Text, textX, currentY-(2*doc.CommentMarkerRadius), 0, &c.Format.FontText)
 		currentY += getMax(c.TextHeight, neededMarkerSpace)
 	}